Reorganized project files & added some scripts.
