+++ format.sh
@@ -1,0 +1,4 @@
+black src
+black tests
+autoflake --in-place --remove-all-unused-imports --remove-unused-variables -r src
+autoflake --in-place --remove-all-unused-imports --remove-unused-variables -r tests
