+++ run.sh
@@ -1,0 +1,1 @@
+python src/ophidian.py
